+++ vm/CPU/core0/gate.h
@@ -3,7 +3,7 @@
 
 class Gate {
       public:
-        int route(long,long,long,long);
+        int route(double,double,double,double);
 
 };
