+++ vm/CPU/core0/lu.h
@@ -1,12 +1,12 @@
 #ifndef LOGIC_UNIT
 #define LOGIC_UNIT
 
-void and_l(long *pckg);
-void or_l(long *pckg);
-void xor_l(long *pckg);
-void not_l(long *pckg);
-void nand_l(long *pckg);
-void nor_l(long *pckg);
-void xnor_l(long *pckg);
+void and_l(double *pckg);
+void or_l(double *pckg);
+void xor_l(double *pckg);
+void not_l(double *pckg);
+void nand_l(double *pckg);
+void nor_l(double *pckg);
+void xnor_l(double *pckg);
 
 #endif
